Feature/95/suhyun (#96)
* fix(): app-api 경로 이름 삭제

* refactor(): 예약 업데이트 1분 ->15분으로 변경

* refactor(): 불필요한 쿼리문 제거

* refactor(): getReservations를 jpql을  사용하여 한방쿼리로 수정

* refactor(): 불필요한 필드 삭제

* refactor(): Reservation 지연 로딩 설정

* refactor(): 불필요한 로직 주석 처리

* refactor(): 각 api의 응답시간 체크를 위한 log 추가

* refactor(): 도착지, 출발지 enumerate(EnumType.STRING)로 변경

* feat(): 결제 완료 후 메세지 큐 구현

* fix(): 예약 취소 검증 로직 오류 수정

* feat(): 결제 완료 후 이메일 전송 기능 구현

* refactor(): 결제 정보 저장 로직 동기처리로 변경 - > 이메일 전송만 비동기처리

* fix(): 주석 변경처리

* feat(): 결제 후처리 비동기 사용, 후처리 후 이메일 전송 메세지 큐를 이용하여 처리

* refactor(): 메서드 분리

* feat(): 이메일 발송 @Async 이용 코드 추가(주석 처리)

* refactor(): 예약 진행 시 비관적인 락 설정

* build(): 이메일 전송 로직 잠시 주석 처리

* build(): 테스트 잠시 주석처리

* build(): 로그인 테스트 잠시 주석처리

* test(): 결제 테스트 수정

* test(): generator 주석처리

* build(): messagemq 주석처리

* build(): rabbitMq 주석처리

* build(): rabbitMq 주석처리

* build(): email관련 주석처리

* fix(): 에러 확인 알림 변경

* fix(): 티켓 관련 로직 주석처리

* fix(): @EntityGraph 적용

지연로딩 사용으로 인해 OSIV OFF로 설정할 시 데이터가 불러와지지 않는 에러 해결

* refactor(): transactional 적용

* feat(): 지연로딩 적용

* refactor(): 예약 동시성 문제 해결

* build(): 메세지큐 관련 주석 원복

* chore(): 예약 후 15분 내 결제 유도 메세지 추가

* refactor(ReservationService): 가독성을 위한 리팩터링 및 메서드 분리

* refactor(reservationResponseDto): 필요 없는 컬럼 삭제

* refactor(): 필요 없는 컬럼 삭제

* refactor(): 필요 없는 코드 삭제

* refactor(): 예약기능 비관적락으로 동시성 제어

* refactor(): 좌석 번호 수정

* refactor(): 티켓에서 출력번호를 결제id->예약 id로 변경

* refactor(): 티켓에서 출력번호를 결제id->예약 id로 변경

* Resolve merge conflict in ReservationServiceImpl.java

+++ src/main/java/com/example/pigonair/domain/reservation/dto/ReservationResponseDto.java
@@ -7,7 +7,7 @@ public record ReservationResponseDto(
 	LocalDateTime departureTime,
 	String origin,
 	String destination,
-	Long seatNumber,
+	int seatNumber,
 	Long price
 
 ) {
